+++ src/main-autoset.js
@@ -1,25 +1,5 @@
+import { get_print_index_by_date } from './get_print_index_by_date.js';
 import { apply_to_buttons } from './apply_to_buttons.js';
-
-const log = function(){
-    const con = window.console;
-    if(COLOR.debug){
-        con.log.apply(null, Array.from( arguments ) );
-    }
-};
-const mod = function(x, m) {
-    return (x % m + m) % m;
-};
-
-const get_print_index_by_date = (COLOR, DATE=new Date()) => {
-    const time = DATE.getTime();
-    const ms = Math.round( time  - COLOR.first_tuesday.getTime() );
-    const week_count = Math.floor( ms  / ( 1000  * 60 * 60 * 24 * 7) );
-    if(COLOR.autoset){
-        const week_delta = week_count * ( COLOR.ascending ? 1 : -1 );
-        return mod(COLOR.first_index + week_delta, COLOR.data.length);
-    }
-    return COLOR.first_index;
-};
 
 const COLOR = {
     autoset: true,